Rejigger how state information is retrieved
In the current previous, `self.state` and `self.player_index` were
recomputed (i.e., retrieved via a JS script call) each time they were
referred to, which is 10s of times per action taken. Beyond being
computationally expensive, this is actually undesirable because it
means that a single action decision is not self-consistent with respect
to the underlying state, which is changing during the computation for
that action. This commit creates a separate `self.state` property and
`self.observe_state()` function. The state observation is the only one
that actually reaches into the JS, and it's called once per decision.