Wrap sending chat in a try/catch/finally
If the sending of a chat message throws an expection, either due to the player id not being available such as with a broken page load or because of an error raised in broadcast, it would cause the page to submit and therefore reload without the participant_id. This would cause the user's chat history to be erased and their session to be converted to a spectator.