Change WebGL primitive from points to triangles
Points cannot be textured, as they're seen by WebGL as, well, points. The closest to what we need would be Quads, but as this isn't a SEGA Dreamcast we can only do lines, triangles, or compositions of those. This changes the point to be formed of two triangles, each taking three adjacent corners of the quad. As this means there are now six calls to the fragment shader per square (one for each vertex), we also increase the number of primitives by a factor of 6.