Multiply the color by a texture when drawing
First, we set up a list of available textures, currently just a 1x1 white texture. Then we modify the shaders to detect how many textures are available and use ECMAScript6's template string feature to generate appropriate data structures and a lookup for the correct texture to use. As GLSL does not allow indexing a uniform array with a varying we create an unrolled loop of comparisons with const ints to determine what index to use. Note also, that varying int is not a supported data type, so we accept varying float and cast it to an int within the fragment shader.

This also introduces a mirror to the layout() function that returns texture coordinates associated with the 6 vertices. The shader does the work of interpolating te inner pixels, but we need to provide anchor points.