Allow experimenters to request identicon textures for players
A fork of the identicons.js identicon creator has been added to generate pixel buffers of identicons, rather than PNG images. We assume no more than 100 are needed, and pregenerate them, pushing them into the uniform array of textures. The value of the identicons is the MD5 hash of the experiment id with the participant id appended. A UUID5 might be a better choice here, but Javascript is sadly lacking in crypto primitives. MD5 was chosen as being simple to implement, and offering a sufficiently long hash and cascading properties.

These identicons are recoverable given the data stored about the experiment, but the set should be approximately unique across playthroughs.

The config variable  determines if identicons should be shown. The default is false.