Fix on macOS
This changes the way the textures are passed to use a single large texture rather
than a variable length uniform array of textures. The selection of textures
is done by modifying the texture coordinates to index the correct slice of
the texture. This fixes the problem on MacOS that causes rendering to
fail if more than 15 identicons are required.

In addition, this reduces the contrast within an identicon, by request
of Jordan. It also uses the size of the box as an argument to the identicon
generator. This makes for crisper identicons at many sizes.