Story 260 - mutable colors should be limited by num_colors
This change removes the hard-coded colours from the frontend Javascript
and populates them from the settings in the backend instead. This causes
the colors being iterated over when changing color to be limited to the set
defined by num_colors, but also allows for updating colors in a single place.