Add spatial chat
This adds a chat_visibility_threshold parameter to GridUniverse which is a float in the range 0..1.
If the sender of a message's dimness is below the threshold the message won't be relayed. The default
is 0.25, meaning users outside the highlighted area will not be able to send messages to the ego
player unless the threshold is reduces.