Add expected end time to bots
This keeps track of when the bot thinks an experiment run will end
and voluntarily stops 30 seconds later, even if it seems the experiment
is still active. This protects against issues where the bot loses
its connection to the server and therefore does not get updates.

It does, however, break future features that may cause the game length
to change dynamically.